requirements.txt
asgiref==3.8.1
brotli==1.2.0
cffi==2.0.0
chardet==5.2.0
cssselect2==0.8.0
Django==5.1
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-filter==24.3
django-rest-framework==0.1.0
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
fonttools==4.60.1
inflection==0.5.1
jpype1==1.5.2
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
reportlab==4.3.1
setuptools==80.9.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tinycss2==1.5.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
